Add data interval template variables for MDSS (#743)
* Add data interval template variables for MDSS

We want each Airflow task to load a specific, known slice of the source
data. These variables will be populated by Airflow for each dag run with
the logical data interval bounds to enable this.

* Update to newer image version

The newer image contains some updates dependencies, which should fix
some vulnerabilities which were detected on the image.